Refactor code: add warehousing_date() method
의류 입고일을 반환하는 메소드를 추가했습니다.
entry_date라는 용어의 entry 부분이 중의적인 의미가 많아 입고일로
인지하기 어려운 관계로 정확한 warehousing이라는 용어를 사용했으며
테스트 코드에서도 입고일을 구하기 위해 연산하고 있기 때문에
코드의 중복 및 실수를 막기 위해 별도 API로 분리합니다.

+++ lib/OpenCloset/Schema/Result/Clothes.pm
@@ -476,40 +476,62 @@ Composing rels: L</order_details> -> order
 
 __PACKAGE__->many_to_many( "orders", "order_details", "order" );
 
-=method rentable_duration
+=method warehousing_date
 
-의류의 입고일로부터 오늘까지의 날 수를 반환합니다. 2014년 12월 17일 재고관리
-시스템 도입시점 이전에 입고된 의류의 경우 이전 대여자료가 존재하지 않기때문에
-입고일을 시스템 도입시점으로 합니다. 만약 입고일이 오늘보다 앞설경우 음수를
-반환합니다.
+의류의 입고일을 반환합니다. 2014년 12월 17일 재고관리 시스템 도입시점 이전에
+입고된 의류의 경우 재고관리 시스템 도입시점을 입고일로 반환합니다.
 
 =cut
 
-sub rentable_duration {
+sub warehousing_date {
     my $self = shift;
 
-    my $start_dt = DateTime->new(
+    #
+    # 입고일: 기증 행위가 생성된 날짜
+    #
+    my $warehousing_dt = $self->donation->create_date->clone;
+    $warehousing_dt->set_time_zone("Asia/Seoul");
+    $warehousing_dt->truncate( to => "day" );
+
+    #
+    # 시스템 도입 시점: 2014년 12월 17일
+    #
+    my $system_start_dt = DateTime->new(
         year      => 2014,
         month     => 12,
         day       => 17,
         time_zone => "Asia/Seoul",
     );
 
-    my $create_dt = $self->donation->create_date->clone;
-    $create_dt->set_time_zone("Asia/Seoul");
-    $create_dt->truncate( to => "day" );
-
-    my $base_dt = $create_dt;
     #
     # 재고관리 시스템 도입시점 이전에 입고된 의류의 경우
     # 입고일을 시스템 도입시점으로 함
     #
-    $base_dt = $start_dt if $create_dt < $start_dt;
+    $warehousing_dt = $system_start_dt if $warehousing_dt < $system_start_dt;
+
+    return $warehousing_dt;
+}
+
+=method rentable_duration
+
+의류의 입고일로부터 오늘까지의 날 수를 반환합니다.
+입고일이 오늘보다 앞설경우 음수를 반환합니다.
+
+=cut
+
+sub rentable_duration {
+    my $self = shift;
+
+    my $base_dt = $self->warehousing_date;
+    return unless $base_dt;
 
     my $now_dt = DateTime->now( time_zone => "Asia/Seoul" );
     $now_dt->truncate( to => "day" );
 
     my $delta = $base_dt->delta_days($now_dt)->in_units("days");
+    #
+    # 입고일이 오늘보다 앞설경우 음수를 반환합니다.
+    #
     $delta = $delta * -1 if $base_dt > $now_dt;
 
     return $delta;
